pyproject.toml
[tool.poetry]
name = "repl_nix_DOMINATORXUSERBOT"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
lottie = "^0.7.0"
wikipedia = "^1.4.0"
barcode = "^1.0.3"
pyfiglet = "^0.8.post1"
Pyrogram = "^2.0.106"
IMDbPY = "^2022.7.9"
"nekos.py" = "^1.1.0"
sickrage = "^10.0.71"
wget = "^3.2"
bs4 = "^0.0.1"
sql-helper = "^0.0.18"
aiohttp = "^3.8.5"
googletrans = "^3.0.0"
patool = "^1.12"
GitPython = "^3.1.35"
numpy = "^1.25.2"
requests = "^2.31.0"
heroku3 = "^5.2.1"
urllib3 = "^2.0.4"
goap = "^0.3.0"
SQLAlchemy = "^2.0.20"
humanize = "^4.8.0"
glitch-this = "^1.0.2"
selenium = "^4.12.0"
emoji = "^2.8.0"
helpers = "^0.2.0"
validators = "^0.22.0"
Telethon = "^1.29.3"
pytz = "^2023.3.post1"
opencv-python = "^4.8.0"
Pillow = "^10.0.0"
youtube_dl = "^2021.12.17"
sql = "^2022.4.0"
search-engine-parser = "^0.6.8"
JustWatch = "^0.5.1"
pySmartDL = "^1.3.4"
nodejsscan = "^3.7"
gTTS = "^2.3.2"
ujson = "^5.8.0"
geopy = "^2.4.0"
utils = "^1.0.1"
html-telegraph-poster = "^0.5.0"
qrcode = "^7.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
